Se cargan los dropdown para filtrar los horarios disponibles

+++ TPC_Brandan_Repolledo/Dominio/Disponibilidad.cs
@@ -8,7 +8,7 @@ namespace Dominio
 {
     public class Disponibilidad
     {
-        public int IdDisponibilidad { get; set; }
+        public long IdDisponibilidad { get; set; }
 
         public Horario Horario { get; set; }
